requirements.txt
Django==4.2.13
django-cors-headers==3.13.0
djangorestframework==3.15.2
django-rest-swagger==2.2.0
drf-yasg==1.20.0